convert markdown logic to re-usable component
